controller: build KubeletConfigHandler for platform
The `makeKCHandlerForPlatform` provides a common handler that can
be used on both HCP and OCP platform.

This way allow us to narrow down the changes in the logic in to
a centeral close-scoped function in the code,
without having the need to branch the logic for HCP and OCP
in other areas in the Reconcile loop.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>