annotation: emergency annotation to enable custom policy
When cluster gets upgrade from 4.1X -> 4.18
the operator removes the MachineConfig
(that contains the custom SELinux policy)
and using the built-in policy instead.

But we still want to have a way to enable the custom SELinux in case that
something breaks during upgrade.

For that we introduce a special annotation that would be present
in the NUMAResourcesOperator CR and would force it to use the
custom (legacy) SELinux.

Once the annotation specified it would apply on all
RTE pods regardless of their association with the
different NodeGroups.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>