api: types: support PoolName option under NodeGroup
This PR introduces a new API option for selecting nodes on which the RTE will be running, exposing the nodes' topologies, as a preamble to supporting TAS on HCP (hosted control planes) known as Hypershift.

On OCP:
 The way that cluster nodes are grouped is by using Machine Config Pools (MCPs); On the operator side, numaresourcesoperator CR defines an MCP selector that behind the scenes uses the node selector on the corresponding MCP and sets it under the RTE daemonset's node selector.

On HCP:
 The MCP term does not exist on hypershift. Instead, the hypershift platform has a node-pool, which essentially groups the nodes based on a specific node-pool label that is added under the node's labels.

This PR proposes to enable additional options for selecting node groups to run the RTE pods that will work on both OCP and HCP platforms. The new API option called `PoolName` does not change or affect how the MCP selector is processed; It provides a new way for OCP to address the nodes by simply setting the name of the MCP as PoolName, which is equivalent to setting the MachineConfigPoolSelector to a label matching the desired MCP. The new field is optional; thus, this solution is backward compatible.

On HCP, PoolName represents the node pool name to which the nodes belong and on which the user desires to operate the operator (RTE daemonset). Tackling HCP platform is yet to be fully supported and it requires additional
modifications.

Notes & restrictions:
- PoolName represents a single string defining the name of the pool, be
  it MCP name on OCP or NodePool name on HCP.
- Only one pool specifier should be specified per node group; more than one will not be tolerated and will cause a degraded state.
- On OCP where both options can be set, the operator state will be degraded if PoolName of one node group and MachineConfigPoolSelector of another node group points to the same MCP.
- Apart from the aforementioned, no extra validations are applied to determine whether nodes of one node group correlate with nodes from another node group. The user takes responsibility for providing nonconflicting nodes per selector group.
- As for MCP selectors, the selected nodes are not validated for whether or not they have the correct machine configuration needed for TAS to be operational.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>