e2e: change MCP wait logic
Now that we're using a built-in SELinux policy,
we do not wait for the MachineConfig creation.

We'll wait for MCP update only when the legacy policy
is used or when we create KubeletConfig for the tests.

In addition, we'll wait for MCP to transition
from updated <-> updating.

The old behavior was to wait for specific MachineConfig
name, but in case of KubeletConfig creation for example
it is not good enough.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>