apply: new ApplyState function supporting deletion
Add support to deleting objects, not just to create or updated them.
After all, "this object must not be present" is a valid desired
state as well.

To minimize risks and API changes, add a new top-level API which
supports deletion, augmenting the old API which was only supporting
create or update; the old API won't gain delete-support, but will
be deprecated when the new one graduates to stable, hopefully
later on in this cycle.

Signed-off-by: Francesco Romani <fromani@redhat.com>