ctrl: status: nrop: explicit diff check
Instead of relying in the chain of helpers to report
if something changed in status, thus warrants a update,
run a full diff of our status explictely.

If something worthy (whose definition depends on
the helper implemented in `pkg/status`) changed,
then we will push a status update.

This is probably slower than relying on the reports from
subfunctions, but simplifies and streamlines the code
significantly.

Signed-off-by: Francesco Romani <fromani@redhat.com>