[release-4.17][manual] api: fix v1alpha1 backward compatibility
Back in 4.16.0 we changed defaults and added options to both
operator and scheduler objects. Due to the sharing of the config
objects between spec and status and due to the fact we still
serve v1alpha1, we may end up with fields values in the status
which are not declared in the API definition.

The most obvious offender is the enum values, but this
actually affects all the values we set explicit defaults
for and we expose in the status.

This is what OLM complained about.
To fix this, we align the status definition from v1 to v1alpha1,
until we deprecate the v1alpha1 and we stop serving it.

xref: https://github.com/operator-framework/operator-lifecycle-manager/blob/master/doc/design/dependency-resolution.md

Manual backport of #1040

Signed-off-by: Francesco Romani <fromani@redhat.com>