tree: reorganize nrovalidator,nrtcacheck, their imports
Our own `nrovalidate`/ validator tooling is useful, alebit little known.
The main problem, however, is
1. first, the fact of the validator logic imports code in `internal`,
   which make it a wrong fit for sitting in `pkg/validator`. Still,
   here it lies.
2. second, the unnecessarily awkward confusion between `pkg/validator`
   and `pkg/validation`.

To streamline things, we move all the logic in a new subtree
`validator`, much like we always had for `rte`.
We take this chance to avoid the sub-tree `pkg`, which doesn't
make much sense in this context. This removes a large annoyance.

With some trivial Makefile (and a lesser but still thorny import
fix in the e2e serial suite) change, we are good to go again
because the main(/only) consumption vehicle for this logic
is the `nrovalidate` tool.

In addition, downgrade the even-less-known `nrtcacheck` tool
to sit in the `tools` hierarchy. This is an internal only change
with no external impacts at all, as this code was already only
providing a command line UI for some packages.

Signed-off-by: Francesco Romani <fromani@redhat.com>