controller: use slice instead of map to conserve consistent order
Using map to store the daemonset of the targeted MCPs turned out to be a
not the best data structure especially when we would like to conserve a
consistent order of the elements that aligns with the rest of the slices
in the operator status.
The problem with map is that it cannot be guaranteed that the elements
will be ordered in the same order they were placed. To fix that, define
a new struct of an MCP and its respective created RTE daemonset and
replace the map with a slice of this struct which ensures that same data
will produce same output every time.

ref: https://groups.google.com/g/golang-nuts/c/YfDxpkI34hY/m/4pktJI2ytusJ

Signed-off-by: Shereen Haj <shajmakh@redhat.com>