api: allow single MCP match per node group by default
Because of historical accident we have a 1:N mapping between NodeGroup and MCPs (MachineConfigPool*s* is a slice!).
One of the key design assumptions in NROP is the 1:1 mapping between NodeGroups and MCPs.

Since we want to conserve the backward compatibility we still need to
support a data structure that can hold multiple MCP matches per tree;
However, since by design the intention was to have a single matching MCP
per node group, make this the default behavior. Degrade the opertor
status if multiple MCPs were found per node group **and** the operator condition is Available
(doesn't point to other failure).

 To re-enable the old behavior permitting multiple pools matches, use the
following annotation: `config.node.openshift-kni.io/multiple-pools-per-tree: enabled`

Experiment the effects of the new default for the next 2
releases. If there are no complaints, we shall remove the multiple MCP match logic.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>