ctrl: update RTE annotations and labels properly
Labels and annotations and rarely set but the problem with the old code was that
new sets are added on top of what exists in the current sets either
update the value or add a new key:value pair. However, the case where
the key:value pair is not found in the new set doesn't make a difference
in the new map state and if the key is found in the current map it will
still exist in the updated one, while the required is to omit it.
Apart from the API validations, the controller doesn't validate the
labels set nor the annotations.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>