mustgather: Dockerfile: use ocp CLI base image
Initially, must-gather for NROP was built by grabbing the latest
openshift-ose-must-gather image, copying the required binaries for
running must gather and override the collection scripts and base that
on rhel-minimal.

While this works, there is no need for extra dependencies on images when
all we need is the openshift clients to gather the must-gather
collection which is available in the CLI image.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>