hack:ci: remove trap
The run-test-install-e2e.sh script is usually running as part of the u/s CI tests.
We don't want to catch the signal in trap because it's
masking the error which makes the test a false positive.

We want to check the error explicitly and fail loud instead,
because on CI there's no bother to revert the cluster back
to its initial state.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>