e2e: mustgather: fix test to verify TAS enabled nodes only
So far the CI worked with the default node labels which is "worker" and
the NodeGroups of the numaresourcesoperator CR we expected to target the
`worker` mcp. However, this isn't always the case and the test should be
smarter to determine which nodes it should look for their folders in the
collected must-gather.

Adjust the step of building the expected CRs paths to include only nodes
that the operator targets.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>