deploy:hypershift: placeholder for deployer interface
Creating a placeholder for implementing the deployer interface
for HyperShift.

The reason we can't implement it for now is because
we need to figure out where is the KUBECONFIG file
for the HyperShift MNG cluster is located and providing it
to the test.

This is needed because we need to create a KubeletConfig ConfigMap
against the MNG cluster and add it under the NodePool spec.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>