sched: always set the leader election parameters
Because of a overlook, we used to set the leader election
params only if replicas > 1 was requested. This left
the key *and default* corner case of replicas=1 with
compiled in defaults, which are questionable at best
and most likely harmful for our use case.

Make sure to always set sane parameters, obviously
taking into account the user desires from the spec
(Replicas field).

Add more logs to make troubleshooting easier

Signed-off-by: Francesco Romani <fromani@redhat.com>