ctrl: update NodeGroupStatus for Hypershift path
The controller reconciliation should reflect the node group statuses
after verifying all components are up and healthy. To preserve backward
compatibility, for OCP we copy the node group config from
MachineConfigPools[].Config to NodeGroupStatus[].Config. For hypershift
we don't have MCP so we copy that info from the NodeGroup[].Config. That
is safe because we know the controller loop will exit if a
node group is misconfigured.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>