e2e: check configmap cleanup
Check that if we're deleting the kubeletconfig
the derived configmap that nrop generated from it
has been removed as well.

This kind of check unique for hypershift,
since we cannot set the configmap owner reference
to be the kubeletconfig configmap because it lies
on a different namespace.

Therefore, we need to validate that our custom cleanup
logic is working.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>