wait: more robust condition check
instead of following suit the MCP status transition
Updated->Updating->Updated, which is racy and relatively fragile,
let's use a stronger condition:

we care about waiting for MCP to be updated with the desired state.
So:
1. we record the current MCP resourceversion
2. we ask for our desired state
3. we wait for the MCP desired state to be `updated` again *but with
   different resourceversion*

this should allow us to capture the conditions which describe our
desired state while waiting for a stable state in a more robust way.

Signed-off-by: Francesco Romani <fromani@redhat.com>