Separated NROP uninstall from install script
Invoking make test-install-e2e currently handles both the installation and uninstallation of NROP and its components.
We propose separating this into two distinct suites, introducing a dedicated script for the uninstallation process.
This separation would be beneficial in scenarios where we need to install and test the operator, upgrade it, and only then proceed with its deletion.

Signed-off-by: Ronny Baturov <rbaturov@redhat.com>