controller: pause reconciliation when annotation is set
The annotation is per CR.
If CR contains the annotation, the operator will pause
and won't reconcile it.

We're are not adding this functionality to the kubelet_controller
becuase it doesn't watch a CR owned by NROP.

Instead it watches for KubeletConfig object, which is not
exclusivly related to NROP so we'll defer this funtionality to
the future and only if there'll be a valid case.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>