objectstate: rte: extract platform-specific state
extract helper to fetch the state from either
NodeGroups or MachineConfigPools into helpers.
The goal here is to isolate this code to make the main
calling functions more targeted and to unblock future
refactoring.

Add helpers to inject parameters.
This helps making the objectstate flavours all consistent
with each other; previously `rte` was the outlier and the
only variant with extra parameters.

This require the constructor function to return a pointer
to a struct (not a struct). Update all the variants for
consistency.

Signed-off-by: Francesco Romani <fromani@redhat.com>
(cherry picked from commit b4964b1fba5aba7853f20ced83876961abdba697)