objstate: rte: factorize code
the helpers we extracted in previous commits still share some state;
pack them in a struct, which makes also nicer to consume them from
the caller site(s), acting almost like a type.

In the (far?) future, when we will move completely to nodegroups
(if ever), this code should be dissolved back in the callers.

Signed-off-by: Francesco Romani <fromani@redhat.com>