buildhelper: report gitbranch, try harder to get version
Report git raw branch name
Also try first to fetch the version from the master location
(Makefile) before to deduce fromthe branch name.
The immediate benefit is we have a much higher chance to always
report a version number, because we must have a version number
in the master location in the Makefile.

Signed-off-by: Francesco Romani <fromani@redhat.com>