serial:config: validate smt alignment is off
Some of the test pods are asking for an odd number of CPUs.
if SMT alignment enabled the test will fail in the middle of the
run and it will requires extra time and effort to understand the
root cause.

Lets avoid that and fail loudly in the beginning of the test
explaining that SMT alignment should be disabled.

The serial suite has a unique characteristics and it should
be running on a dedicated environment,
so although smt alignment should usually be enable
for the test purposes it make a complete sense to require
its disabelement.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>