e2e: classifying tests using labels
The label package contains a list of labels that can be used in
the e2e and serial tests to indicate and point certain behavior or characteristics
of the various tests.

Those can be filtered/focused by ginkgo before the test runs.

Example:
`ginkgo --label-filter="!slow"` means avoid any specs labelled `slow`

The types defined in the package are type aliases and not new defined
types. This is done in order to avoid casting every time the labels are
being used.

This package allow greater readability, prevent typos (what can easily
happen while using plain-text) and
a common place to hold and documents every label that is being
used in the tests.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>