pkg: use go-cmp only in tests
Similarly to https://github.com/kubernetes/kubernetes/issues/104821 , we
use the go-cmp package in our non-test code. Problem is: this package
reserves the right to panic, which is something we cannot accept in out
non-test code. We need to replace to similar code which guarantee not to
panic.

Replace go-cmp.Equal with reflect.DeepEqual in non-test code.
more on reflect.DeepEqual: https://pkg.go.dev/reflect#DeepEqual.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>