remove GinkgoHelper() for better debugging
TBD
see ginkgohelper description.

Problem:
the reboot test 47674 is failing but with unclear debugging logs which
point to somthing went wrong in an early step but reports that at tear
down without the full stacktrace which is poiting to the function that
calls GinkgoHelper.

What do you want?
Be able to track down the root cause of the test.

How are you fixing this?
remove GinkgoHelper to enable logging the exact failing line.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>