remove GinkgoHelper() for better debugging
The reboot test 47674 was reporting a failure while waiting for MCP to get updated but the pointer to the failing line was a line within the test spec (It()), without extra details on where specifically the failure happened within the helper function, in which we call GinkgoHelper(). Due to insufficient logging, it was even harder to trace where the failure root is (more clear logging will be added to avoid this as we advance).
What GinkgoHelper does is that wherever it is added to a function, in case a failure happen during the execution of the function body, ginkgo will exclude this function from the stack trace, a thing that we highly need especially for the serial tests and reboot ones to be able to troubleshoot issues effectively.From the beginning adding GinkgoHelper() call to the update function was mistake because it isn't a proper helper. We need a better difinetion of helper function in the suite, which will follow in later PRs but for now remove this call to enable better troubleshooting.

The purpose of this PR is not to fix any of the tests failures but to enable logging the exact failing line even if in helper functions, after all these are part of the tests and they too perform some unique verifications, and if the internal verifications fail the test is very likely to fail.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>