linter: enable copyloopvar check and fix instances
While the logic behind this linter is understood, one needs to be careful
to apply it especially when having goroutines within the loop because they execute
asynchronously. To guarantee that the variables used in these routines
don't interfere and still solve the linter error, we can either move the copy into the goroutine scope or simply avoid shadowing by getting the value by the index in range. In this commit do the latter.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>