objectstate:setdefault: set default values to objects in desired state
Many of the `Update` calls done by the operator
are changing nothing.

This is happening because the API server defaults some of the fields
while our code leave them empty.

This result in a diff in the comparison function and leads to redundant calls.

This package purpose it to set the default values in our objects similar to what
the API server is doing.

The code is based on upstream code that is
not in staging so we can't import it unfortunately.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>