linter: enable errcheck and fix instances
The `errcheck` linter is especially important to avoid ignoring errors
that might be or cause critical bugs. Enable this linter and fix the
instances. Note that sometimes ignoring the errors is intentional to
for example if the error isn't related to the flow of the functionality
that is intended to run like reporting to ginkgo writer; However we
don't want to ignore all errors when writing data using `fmt.Fprintf()`
or similar because it depends on whether the writer is being used and
will later be analyzed during the executed code.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>