serial:47674: fix mcp anticipated updates
With the latest selinux updates, NROP will no longer require to reboot
the nodes on CR updates unless a node group annottaion is enabled. The
new selinux policy is default starting 4.18. However this wasn't
backported to older versions so in this commit we adapt this change
based on the existance of the annotation and the openshift version.
This commit fixes the expected bahavior per each case and adds missing
waiting loops for MCP updating conndition.

The test passed with default config on 4.18 and on older versions.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>