serial: rewrite test 47674
Use simpler approach considering this is not checking the changes e2e
Previuosly the test was creating a new MCP and managed the nodes labels
which caused additional reboot cycles especially for versions <4.18.
In this new version of the test we use existing MCP (always the master
in d/s ci) and switch to use that in the NROP CR. The test is only
checking that the changes in the CR are reflected in the DS and not
actually e2e of checking RTE pods deployed or not so we are using that
aspect.
Considering several bugs were discovered in this area and need more
coverage this commit is only allowing us to have same old coverage with
simpler test steps and less maintainance of it in the future.

PS: this was tested only on defaul 4.18; running it on <4.18 and
4.18+custom policy is yet to be done.

How? TBD

Signed-off-by: Shereen Haj <shajmakh@redhat.com>