deps: bump x/oauth2 to 0.26
This was prompted by GO-2025-3488 but we are NOT AFFECTED
by that CVE because we don't use, thus we don't vendor,
the relevant code at all.

Still, we bump the dep to the last version which is compatible
with the golang version we use in this branch, to be proactive.

Signed-off-by: Francesco Romani <fromani@redhat.com>