bump deps: glog, x/oauth2
bump glog dep to proactive remove possible vulnerabilities. Note our production code is NOT AFFECTED by GO-2025-3372, because we do not use in production code, but only in test code.
Also bump x/oauth2 dep. This was prompted by GO-2025-3488 but we are NOT AFFECTED by that CVE because we don't use, thus we don't vendor, the relevant code at all.

Still, we bump the deps to the last version which is compatible with the golang version we use in this branch, to be proactive.

Signed-off-by: Francesco Romani <fromani@redhat.com>