e2e: tolerate devices in reboot tests
Problem:
Reboot tests fails frequently on teardown when comparing the quantaty of
the device resources. The root cause is from the sample-device plugin
but haven't yet fully investigated considering we **want** to move to
real devices/ sriov simulation and due to capacity.

What do you want?
Make the serial suite less noisy due to to known
issue, until we fix the root cause or implement the alternative:
https://issues.redhat.com/browse/CNF-12824

How are you fixing the problem in this PR?
similarly to memory resources deviation for reboot tests, allow devices
quantity to change after reboot but under the condition that the ratios
are the same before and after the test in the context of allocatables
and availables; That means capacities of the same device may change
before vs after the test, but unequal resources consumption is not
tolerated.

How did you test the change?
On a cluster with sample devices simulate the reboot scenario by running
an empty test with only a time.Sleep() command, once the sleep period
starts, update the devices count manually on the cluster (reapply CM and
restart the DS). The teardown should reflect that the deviation is
noticed but not failing on that.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>