serial: enhance MCP update logs using identifier
When GinkgoHelper() was present (f56658a)
It was hard to follow the logs and understand which step failed while
waiting for MCP update. Adding unique identifiers to each call helps
track the start and end of each call and enhanced tracking down the root
cause. To guarantee using a unique identifier, we use `time.Now()` as a string.
Additionally report config values and expectations of the call.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>