controller: track rte configmap state
Kubelet controller is responsible to track `KubeletConfig` object and
generate a `ConfigMap` data from it that will be served for RTE.

NROP controller is responsible to track the `ConfigMap` creation,
feed it to RTE and restart RTE in case the `ConfigMap` data has changed.

The later is broken, and this commit is fixing it.

In addition the `ConfigMap` should be per `DaemonSet`/`NodeGroup`, while
previously it was common at the CR level which is wrong.

Because cluster might have multiple node groups which are configured
slightly different.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>