serial: write test 47674 replacement
The previous version was fragile and maintained several times already.
This fact means that the test should be rewritten and once the known bugs
are fixed, new tests will be added to cover the e2e full scenario.

The test goal is to verify that adjustments to NROP CR are done as
expected and reflected in the RTE daemonsets. Previously, this was done
via creating a new MCP for the test and associating to it one of the
worker nodes that were part of the initial NROP node groups. This was
causing a reboot when running on a cluster that is:
a. < 4.18 or
b. >= 4.18 and has the custom RTE selinux policy enabled

In the new version of the test, we keep checking the same functionality
while completely avoiding reboot. That is acheived by creating the new MCP,
keeping it empty (no nodes are associated to it) and appending it as a
second node group's selector. Then checks on the daemonsets are
performed and the only cleanup would be deleting the test mcp.

We keep the old version still because the new one will be skipped from
running on < 4.18 due to https://issues.redhat.com/browse/OCPBUGS-52958.
Once that is fixed, we should settle to the new version only.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>