cleanup: remove hash computation from shceduler's deployment
In contrary to nrop controller where the configmap is generated by a different controller,
on the sheduler side we're fully control the configmap which is static
and not configurable.

We do change the API of one of the functions, but as of now,
we're not aware of anyone using it but this flow.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>