tools: sortimports
A tool to automate the sorting of imports for our files in the project.
The tool expect a one or more  absolute paths on which it will iterate recursevly
sorting the import according to the following scheme:
 1. Standard libraries
 2. "k8s.io" imports
 3. "sigs.k8s.io" imports
 4. Other third-party libraries
 5. "github.com/openshift" imports
 6. "github.com/openshift-kni" imports

The tool can also be given a list of directory to skip, vendor
directory for example.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>