serial: add `openshift` labels
The suite is running on hypershift and some tests that involves MCP
checks fails because MCO is not supported on HCP. Mark these tests with
the `openshift` label so they can be skipped when running the suite on
hypershift for now.
Note that part of these tests with slight adjustments can run on both
platforms; adapting this should be handled in later PRs.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>