deps: e2e: remove dep to performanceprofile API
Depending on the performanceprofile creates an unnecessary dependency
between NROP and NTO. We have 2 tier2 e2e tests which require the
performance profile API types to run, which can and should
be reimplemented using the dynamic client to get rid of the dep.

We will reimplement later. We remove the dep to enable the version bump.

Signed-off-by: Francesco Romani <fromani@redhat.com>