deps: mco: stop IsMachineConfigPoolConditionFalse
the helper is no longer available after the merge of the MCO APIs
in the openshift/api repo.
There is likely be a better way to accomplish this matching,
but until we manage to review the new packages and better
integrate, switch to this stop-gap solution which is meant
to cause no change in behavior.

Signed-off-by: Francesco Romani <fromani@redhat.com>