kubeletcontroller: nrop predicate
Kubelet controller tracks KubeletConfig objects on the system and creates a configmap
data for RTE from them.

There's a subtle bug when a new KubeletConfig gets created on the system,
the controller cannot create the configmap data if a new nodeGroup
was not already configured in the nrop object:
https://github.com/openshift-kni/numaresources-operator/blob/597d9df3ffac7b084fefef8192f28c414c9f3087/internal/controller/kubeletconfig_controller.go#L239

Hence, we should watch for nrop CR changes under the nodeGroup and
retrigger the recociliation loop.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>