internal: ensure total memory resources are equal on node level
In
https://github.com/openshift-kni/numaresources-operator/commit/4f849a40fc669180f036b621f2cfb5c85d89f245
a deviation of the memory amount on NUMA level was added to allow more
relaxed check for tests that involved reboot. This is correct but still
need to verify that the total memory resources on node level is
maintained after the reboot.
Add a verification step to verify that the total memory resources per
compute node is same before and after, and add unit tests coverage.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>