e2e: serial: remove unnecessary mcp update for reboot tests
As NROP CR deletion and creation are not necessary for versions >= 4.18 since it causes tests 75854 and 75855 to fail unintentionally.
The changes here removes the wait for mcp updates and in addition ensures the before each function checks that the RTE pods are deleted properly.
In addition to preserve the old behavior for versions <= 4.17 we're checking to see if the custom policysupport is enabled which then a waitForMcpUpdate is necessary.

Signed-off-by: Roy Shemtov <rshemtov@redhat.com>