schedulercontroller: compute replicas
Adjust the computation function so it will detect replicas number
only when it's not explicitly set by the user.

In addition the default behavior is different now:
HA is enabled by default (previously it was not)
and replicas number is equal to the number of control-plane nodes.

In case the user wants to disable HA (retaining the old behavior)
he needs to explicitly set `spec.Replicas` to 1.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>