e2e: serial: partial rewrite of test 47615
Test 47615 have a precise requirement about resource availability
such as an exact and known number of pods fit on nodes.
Sizing is critical for test, but that was kinda lost in the
updates.

Restore the sizing, add comments to better express the intent
and along the way use more modern constructs to (hopefully) improve
the test reliability.

Signed-off-by: Francesco Romani <fromani@redhat.com>