e2e: serial: schedcache: clarify and fix filterAnyZoneProvidingResourcesAtMost
Some scheduler cache e2e tests were using, and were the sole consumers,
of the filterAnyZoneProvidingResourcesAtMost.
The function is unfortunately confusing because it checks a very
specific thing:
- filter in (keeps only) nodes which have zones
- which provide a given resources to fullfil a request (at least)
- but the count of these zones should be capped till a limit.

In other words, the function wants to answer questions like
"are there nodes which have X resources in at most N numa zones?"
e.g.
"are there nodes which have 8 devices in at most 1 numa zone?"

- PASS: if a node has, like, 10 devices attached to either,
  but not both, numa  zone
- FAIL: if a node has, 6 devices attached to any (or both)
  numa zone
- FAIL: if a node has 8 devices attached to both nodes

and so on.

The function had also a bug: its main work horse callee claimed
in its name it was checking a NUMA zone was providing up to a given
amount of resources (AtMost) while it was checking the opposite,
if the NUMA zone was providing enough of a resource (AtLeast).

In this commit we fix all the above.
NOTE it is possible the test worked by accident and we are now
actually breaking it, but at least now the code is more consistent
and less obscure.

Also fix related typos along the way.

Signed-off-by: Francesco Romani <fromani@redhat.com>